+++ NewDawn/NewDawn/ViewControllers/RegistrationViewControllers/Profile_DraftFinal.swift
@@ -33,7 +33,7 @@ class Profile_DraftFinal: UIViewController {
         // TODO: Send all info to backend and go to profile page
         let activityIndicator = self.prepareActivityIndicator()
         
-        let request = createRegistrationRequest()
+        let request = EditProfileUtil.createRegistrationRequest()
         
         if request == nil {
             return
@@ -52,8 +52,8 @@ class Profile_DraftFinal: UIViewController {
                         "caption": single_image["caption"]!,
                         "user": single_image["user_uri"]!
                         ] as [String: Any]
-                    let img_name = self.MD5(String(single_image["user_id"] as! Int) + String(single_image["order"] as! Int))! + ".jpeg"
-                    self.photoUploader(photo: single_img as! UIImage, filename: img_name, parameters: single_params){ success in
+                    let img_name = String.MD5(String(single_image["user_id"] as! Int) + String(single_image["order"] as! Int))! + ".jpeg"
+                    ImageUtil.photoUploader(photo: single_img as! UIImage, filename: img_name, parameters: single_params){ success in
                         print("image upload \(success)")}
                 }
             }
@@ -64,165 +64,11 @@ class Profile_DraftFinal: UIViewController {
         
     }
     
-    func createRegistrationRequest() -> URLRequest? {
-        
-        var url = getURL(path: "register/")
-        var httpMethod: String;
-        let register_info: [String: Any] = getUserInputInfo()
-        let user_id = LoginUserUtil.getLoginUserId()
-        if user_id != nil && user_id != 1{
-            let user_id_url = String(user_id!) + "/"
-            httpMethod = "PUT"
-            url = url.appendingPathComponent(user_id_url)
-        }else{
-            httpMethod = "POST"
-        }
-        var request = URLRequest(url:url)
-        request.httpMethod = httpMethod
-        request.addValue("application/json", forHTTPHeaderField: "Content-Type")
-        request.addValue("application/json", forHTTPHeaderField: "Accept")
-        do {
-            request.httpBody = try JSONSerialization.data(withJSONObject: register_info, options: .prettyPrinted)
-        } catch let error {
-            // TODO: Sepearte error messages into engineer's and user's
-            print(error.localizedDescription)
-            self.displayMessage(userMessage: "Registration Request Creation Failed", dismiss: false)
-            return nil
-        }
-        return request
-    }
-    
-    func getUserInputInfo() -> [String: Any]{
-        
-        let user_data = getUserData()
-        let account_data = getAccountData()
-        let profile_data = getProfileData()
-//        let visibility_data = []
-        
-        let question_answer_data = getQuestionAnswersData()
-        
-        // TODO: better handle username and password
-        let random_id = UUID().uuidString
-        let pesudo_data = [
-            "username": LocalStorageUtil.localReadKeyValue(key: PHONE_NUMBER) ?? random_id,
-            "password":random_id
-        ]
-        
-        var register_data = [String: Any]()
-        register_data += user_data as! Dictionary<String, String>
-        register_data += account_data as! Dictionary<String, String>
-        register_data += profile_data
-        register_data += question_answer_data
-        register_data += pesudo_data
-        return register_data
-    }
-    
     func readRegistrationResponse(parseJSON: NSDictionary) -> Void {
         let msg = parseJSON["success"] as? Bool
         if msg == false {
             self.displayMessage(userMessage: "Registration Data Failed to Send")
             return
-        }
-    }
-    
-    func getUserData() -> [String: Any]{
-        return  [
-            FIRSTNAME: localReadKeyValue(key: FIRSTNAME)!,
-            LASTNAME: localReadKeyValue(key: LASTNAME)!,
-        ]
-    }
-    
-    func getAccountData() -> [String: Any] {
-        return [
-            "birthday":_birthday_str_handler(birthday: localReadKeyValue(key: BIRTHDAY) as! String),
-            "gender":localReadKeyValue(key: GENDER)!,
-        ]
-    }
-    
-    func getProfileData() -> [String: Any] {
-        return [
-            "height": _height_num_handler(height: localReadKeyValue(key: HEIGHT) as! String),
-            "hometown":localReadKeyValue(key: HOMETOWN) ?? UNKNOWN,
-            "school":localReadKeyValue(key: SCHOOL) ?? UNKNOWN,
-            "degree":localReadKeyValue(key: DEGREE) ?? UNKNOWN,
-            "job_title":localReadKeyValue(key: JOBTITLE) ?? UNKNOWN,
-            "employer":localReadKeyValue(key: WORKPLACE) ?? UNKNOWN,
-            "drink":localReadKeyValue(key: DRINK) ?? UNKNOWN,
-            "smoke":localReadKeyValue(key: SMOKE) ?? UNKNOWN,
-        ]
-    }
-    
-    
-    // Transform mm/dd/yy to yyyy/mm/dd
-    func _birthday_str_handler(birthday: String) -> String{
-        return birthday.replacingOccurrences(of: "/", with: "-")
-    }
-    
-    // Transform string to int
-    func _height_num_handler(height: String) -> Int{
-        if height.prefix(1) == "<"{
-            return 139
-        }
-        return Int(height.prefix(3))!
-    }
-    
-    // Transform QuestionAnswer struct to array of dicts
-    func _question_answer_handler(existed_question_answers: Array<QuestionAnswer>) -> Array<[String: Any]>{
-        var question_answer_array = [[String: Any]]()
-        var num_count = 1;
-        for question_answer in existed_question_answers{
-            let holder = [
-                "question": question_answer.question.question,
-                "answer": question_answer.answer,
-                "order": num_count
-                ] as [String : Any]
-            num_count += 1
-            question_answer_array.append(holder)
-        }
-        return question_answer_array
-    }
-    
-    func getQuestionAnswersData() -> [String: Any] {
-        if let existed_question_answers: Array<QuestionAnswer> = localReadKeyValueStruct(key: QUESTION_ANSWERS) {
-            let question_answer_array = _question_answer_handler(existed_question_answers: existed_question_answers)
-            return ["answer_question": question_answer_array]
-        }
-        return ["answer_question": [[String: Any]]()]
-    }
-    
-    // A helper function to upload image
-    func photoUploader(photo: UIImage, filename: String, parameters: [String: Any], completion: @escaping (Bool) -> Void) {
-        
-        let imageData = photo.jpegData(compressionQuality: 1)
-        
-//        guard let authToken = Locksmith.loadDataForUserAccount(userAccount: "userToken")?["token"] as? String else {
-//            return
-//        }
-        
-        let headers: HTTPHeaders = [
-            "Content-Type": "application/json",
-            "Accept": "application/json"
-        ]
-        var url: URLRequest?
-        
-        let image_upload_url = getURL(path: "image/")
-        
-        do {
-            url = try URLRequest(url: image_upload_url, method: .post, headers: headers)
-        } catch {
-            print("Error")
-        }
-
-        let data = try! JSONSerialization.data(withJSONObject: parameters)
-
-        
-        if let url = url {
-            upload(multipartFormData: { (mpd) in
-                mpd.append(imageData!, withName: "media", fileName: filename, mimeType: "image/jpeg")
-                mpd.append(data, withName: "data")
-            }, with: url, encodingCompletion: { (success) in
-                debugPrint(success)
-            })
         }
     }
     
@@ -232,21 +78,6 @@ class Profile_DraftFinal: UIViewController {
         if let user_id = register_response["id"] as? Int, let token = register_response["token"] as? String {
             _ = LoginUserUtil.saveLoginUserId(user_id: user_id)
             _ = LoginUserUtil.saveAccessToken(token: token)
-        }
-    }
-    
-    func MD5(_ string: String) -> String? {
-        let length = Int(CC_MD5_DIGEST_LENGTH)
-        var digest = [UInt8](repeating: 0, count: length)
-        
-        if let d = string.data(using: String.Encoding.utf8) {
-            _ = d.withUnsafeBytes { (body: UnsafePointer<UInt8>) in
-                CC_MD5(body, CC_LONG(d.count), &digest)
-            }
-        }
-        
-        return (0..<length).reduce("") {
-            $0 + String(format: "%02x", digest[$1])
         }
     }
