+++ NewDawn/NewDawn/ViewControllers/SettingPageViewControllers/SettingPageViewController.swift
@@ -64,13 +64,23 @@ class SettingPageViewController: UIViewController {
                     let user_hometown = user_profile!.hometown
                     self.NameAgeText.text = user_firstname + ", " + String(user_age)
                     self.HomeTownText.text = user_hometown
-                    if user_profile!.mainImages.isEmpty == false {
-                        self.profileImage.downloaded(from:
-                            self.profileImage.getURL(path: user_profile!.mainImages[0].image_url))
+                    if let images_count = LocalStorageUtil.localReadKeyValue(key: "ImagesCount"){
+                        if user_profile!.mainImages.count == images_count as! Int {
+                            self.profileImage.downloaded(from:
+                                self.profileImage.getURL(path: user_profile!.mainImages[0].image_url))
+                            EditProfileTabelViewController.downloadOverwriteLocalImages(profile: user_profile!)
+                            EditProfileTabelViewController.downloadOverwriteLocalInfo(profile: user_profile!)
+                        }else{
+                            self.viewDidLoad()
+                        }
+                    }else{
+                        if user_profile!.mainImages.isEmpty == false {
+                            self.profileImage.downloaded(from:
+                                self.profileImage.getURL(path: user_profile!.mainImages[0].image_url))
+                            EditProfileTabelViewController.downloadOverwriteLocalImages(profile: user_profile!)
+                            EditProfileTabelViewController.downloadOverwriteLocalInfo(profile: user_profile!)
+                        }
                     }
-                    
-                    EditProfileTabelViewController.downloadOverwriteLocalImages(profile: user_profile!)
-                    EditProfileTabelViewController.downloadOverwriteLocalInfo(profile: user_profile!)
                 }
             }
         }
