+++ NewDawn/NewDawn/ViewControllers/ViewController.swift
@@ -7,6 +7,7 @@
 //
 
 import UIKit
+import CommonCrypto
 
 // Below is a library for common used code
 // applied to ALL UI VIEWS
@@ -337,5 +338,22 @@ extension UITextField {
         self.layer.shadowOffset = CGSize(width: 0.0, height: 1.0)
         self.layer.shadowOpacity = 1.0
         self.layer.shadowRadius = 0.0
+    }
+}
+
+extension String{
+    static func MD5(_ string: String) -> String? {
+        let length = Int(CC_MD5_DIGEST_LENGTH)
+        var digest = [UInt8](repeating: 0, count: length)
+        
+        if let d = string.data(using: String.Encoding.utf8) {
+            _ = d.withUnsafeBytes { (body: UnsafePointer<UInt8>) in
+                CC_MD5(body, CC_LONG(d.count), &digest)
+            }
+        }
+        
+        return (0..<length).reduce("") {
+            $0 + String(format: "%02x", digest[$1])
+        }
     }
 }
